+++ src/ontolog/cli/graph/__init__.py
@@ -1,5 +1,0 @@
-"""Graph CLI package."""
-
-from ontolog.cli.graph.commands import graph
-
-__all__ = ["graph"]